+++ relay-server/src/services/outcome.rs
@@ -450,6 +450,9 @@ pub enum DiscardReason {
     InvalidReplayRecordingEvent,
     InvalidReplayVideoEvent,
 
+    /// (Relay) The event's timestamp was too far in the future.
+    DateInTheFuture,
+
     /// (Relay) Profiling related discard reasons
     Profiling(&'static str),
 
@@ -503,6 +506,7 @@ impl DiscardReason {
             DiscardReason::Profiling(reason) => reason,
             DiscardReason::InvalidSpan => "invalid_span",
             DiscardReason::FeatureDisabled(_) => "feature_disabled",
+            DiscardReason::DateInTheFuture => "date_in_the_future",
         }
     }
 }