+++ relay-server/src/services/spooler/mod.rs
@@ -29,12 +29,7 @@
 //!
 //! Current on-disk spool implementation uses SQLite as a storage.
 
-use std::collections::{BTreeMap, BTreeSet};
-use std::error::Error;
-use std::path::Path;
-use std::pin::pin;
-use std::sync::Arc;
-
+use chrono::{DateTime, Utc};
 use futures::stream::{self, StreamExt};
 use hashbrown::HashSet;
 use relay_base_schema::project::{ParseProjectKeyError, ProjectKey};
@@ -48,11 +43,15 @@ use sqlx::sqlite::{
     SqliteSynchronous,
 };
 use sqlx::{Pool, Row, Sqlite};
+use std::collections::{BTreeMap, BTreeSet};
+use std::error::Error;
+use std::path::Path;
+use std::pin::pin;
+use std::sync::Arc;
 use tokio::fs::DirBuilder;
 use tokio::sync::mpsc;
 
 use crate::envelope::{Envelope, EnvelopeError};
-use crate::extractors::StartTime;
 use crate::services::outcome::TrackOutcome;
 use crate::services::processor::ProcessingGroup;
 use crate::services::projects::cache::legacy::{ProjectCache, RefreshIndexCache, UpdateSpoolIndex};
@@ -492,7 +491,7 @@ impl OnDisk {
         let received_at: i64 = row
             .try_get("received_at")
             .map_err(BufferError::FetchFailed)?;
-        let start_time = StartTime::from_timestamp_millis(received_at as u64);
+        let received_at = DateTime::from_timestamp_millis(received_at).unwrap_or(Utc::now());
         let own_key: &str = row.try_get("own_key").map_err(BufferError::FetchFailed)?;
         let sampling_key: &str = row
             .try_get("sampling_key")
@@ -503,7 +502,7 @@ impl OnDisk {
                 .map_err(BufferError::ParseProjectKeyFailed)?,
         };
 
-        envelope.set_start_time(start_time.into_inner());
+        envelope.set_received_at(received_at);
 
         let envelopes: Result<Vec<_>, BufferError> = ProcessingGroup::split_envelope(*envelope)
             .into_iter()
@@ -1313,6 +1312,7 @@ impl Drop for BufferService {
 
 #[cfg(test)]
 mod tests {
+    use chrono::Utc;
     use insta::assert_debug_snapshot;
     use rand::Rng;
     use relay_test::mock_service;
@@ -1410,7 +1410,7 @@ mod tests {
             };
 
             let envelope = empty_managed_envelope();
-            let start_time_sent = envelope.start_time();
+            let received_at_sent = envelope.received_at();
             addr.send(Enqueue {
                 key,
                 value: envelope,
@@ -1430,15 +1430,12 @@ mod tests {
             });
 
             let UnspooledEnvelope { managed_envelope } = rx.recv().await.unwrap();
-            let start_time_received = managed_envelope.envelope().meta().start_time();
+            let received_at_received = managed_envelope.received_at();
 
             // Check if the original start time elapsed to the same second as the restored one.
-            //
-            // Using `.as_secs_f64()` to get the nanos fraction as well and then round it up to get
-            // similar number of seconds if one of the instants runs forward a bit.
             assert_eq!(
-                start_time_received.elapsed().as_secs_f64().round(),
-                start_time_sent.elapsed().as_secs_f64().round()
+                received_at_received.timestamp_millis(),
+                received_at_sent.timestamp_millis()
             );
         }
     }
@@ -1585,7 +1582,7 @@ mod tests {
                 .await
                 .unwrap();
 
-        let now = chrono::Utc::now().timestamp_millis();
+        let now = Utc::now().timestamp_millis();
         let result = sqlx::query(&format!(
             r#"INSERT INTO
                     envelopes (received_at, own_key, sampling_key, envelope)